:recycle: (Entity\Traits) Refactoring using SoftDeletable trait
