+++ src/Entity/Traits/TimestampsTrait.php
@@ -1,0 +1,51 @@
+<?php
+
+namespace App\Entity\Traits;
+
+use DateTimeInterface;
+use Doctrine\ORM\Mapping as ORM;
+use Symfony\Component\Validator\Constraints as Assert;
+
+/**
+ * A trait that entities can use to have timestamps. It adds a `createdAt` and an `updatedAt` property with their getters and setters.
+ */
+trait TimestampsTrait
+{
+    /**
+     * @ORM\Column(type="datetime")
+     *
+     * @Assert\Type("\DateTimeInterface")
+     */
+    private $createdAt;
+
+    /**
+     * @ORM\Column(type="datetime")
+     *
+     * @Assert\Type("\DateTimeInterface")
+     */
+    private $updatedAt;
+
+    public function getCreatedAt(): ?DateTimeInterface
+    {
+        return $this->createdAt;
+    }
+
+    public function setCreatedAt(DateTimeInterface $createdAt): self
+    {
+        $this->createdAt = $createdAt;
+
+        return $this;
+    }
+
+    public function getUpdatedAt(): ?DateTimeInterface
+    {
+        return $this->updatedAt;
+    }
+
+    public function setUpdatedAt(DateTimeInterface $updatedAt): self
+    {
+        $this->updatedAt = $updatedAt;
+
+        return $this;
+    }
+}
